+++ src/components/constructor-ingredient/constructor-ingredient.module.css
@@ -7,3 +7,7 @@
   top: 25px;
   left: -35px;
 }
+
+.ingredient {
+  display: flex;
+}
