+++ src/components/modal/modal.module.css
@@ -1,6 +1,5 @@
 .modal {
   width: 720px;
-  height: 100%;
   box-sizing: border-box;
   position: fixed;
   top: 50%;
