+++ src/components/order-details/order-details.jsx
@@ -1,26 +1,26 @@
 import clsx from 'clsx';
 import styles from './order-details.module.css';
 import done from '../../images/done.png'
-import Loader from "../loader/loader";
+import Loader from '../loader/loader';
 
-const OrderDetails = ({ order }) => {
+const OrderDetails = ({order}) => {
   return (
-    <div className={clsx(styles.order, 'pb-30')}>
-      { order && (
+    <div className={styles.order}>
+      {order && (
         <>
-          <p className={clsx(styles.number, 'text_type_digits-large pt-1 mt-1 mb-10')}>{ order.order.number }</p>
+          <p className={clsx(styles.number, 'text_type_digits-large mt-1 mb-10')}>{order.order.number}</p>
           <p className={'text_type_main-medium mt-5 mb-15'}>идентификатор заказа</p>
-          <img src={done} alt='заказ принят' />
+          <img src={done} alt='заказ принят'/>
           <div className='text_type_main-small mt-15 mb-2'>Ваш заказ начали готовить</div>
           <div className='text_type_main-small text_color_inactive'>Дождитесь готовности на орбитальной станции</div>
         </>
-      ) }
-      { !order && (
+      )}
+      {!order && (
         <div className='m-30'>
-          <p className='mb-20 text text_color_inactive text_type_main-medium'>Подсчитываем стоимость и формируем заказ</p>
-          <Loader />
+          <p className='mb-20 text text_color_inactive text_type_main-medium'>Формируем заказ</p>
+          <Loader/>
         </div>
-      ) }
+      )}
     </div>
   );
 };
