+++ src/services/actions/constants.js
@@ -16,5 +16,4 @@ export const ADD_BUN = 'ADD_BUN';
 export const ADD_INGREDIENT = 'ADD_INGREDIENT';
 export const REPLACE_INGREDIENT = 'REPLACE_INGREDIENT';
 export const DELETE_INGREDIENT = 'DELETE_INGREDIENT';
-export const CALCULATE_PRICE = 'CALCULATE_PRICE';
 export const RESET_STATE = 'RESET_STATE';
