+++ src/services/actions/get-details.js
@@ -1,0 +1,11 @@
+import {   
+  GET_DETAILS_REQUEST
+} from './constants';
+
+
+export function getDetails(ingredient) {
+  return {
+    type: GET_DETAILS_REQUEST,
+    details: ingredient
+  }
+}
