+++ src/services/store.js
@@ -1,6 +1,6 @@
-import { legacy_createStore as createStore, applyMiddleware, compose} from 'redux';
+import {legacy_createStore as createStore, applyMiddleware, compose} from 'redux';
 import thunk from 'redux-thunk';
-import { rootReducer } from './reducers/reducers';
+import {rootReducer} from './reducers/root-reducer';
 
 const composeEnhancers =
   typeof window === 'object' && window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__
